# Conflicts:
#	src/test/java/com/catalyst/collector/services/impl/CollectionsServiceImplTest.java

+++ src/main/java/com/catalyst/collector/services/impl/CollectionsServiceImpl.java
@@ -146,7 +146,7 @@ public class CollectionsServiceImpl implements CollectionsService {
 
     @Override
     public boolean addKeyword(Keyword keyword) {
-        if (keyword.getKeyword().length() < 1 || keyword.getKeyword().length() > 255)
+        if (!keyword.isValid())
             return false;
 
         collectionsDao.addKeyword(keyword);
@@ -155,7 +155,7 @@ public class CollectionsServiceImpl implements CollectionsService {
 
     @Override
     public boolean updateKeyword(Keyword keyword) {
-        if (keyword.getKeyword() == null || keyword.getKeyword().equals("") || keyword.getKeyword().length() > 255)
+        if (!keyword.isValid())
             return false;
         collectionsDao.updateKeyword(keyword);
         return true;
@@ -163,6 +163,8 @@ public class CollectionsServiceImpl implements CollectionsService {
 
     @Override
     public void removeKeyword(Integer id) {
+        if (id < 0)
+            return;
         collectionsDao.removeKeyword(id);
     }
